Minor edits to files, added application props
Minor edits are removing course printing duplication, adjusting the tests, and changing doubles to int

+++ BE-Road2CS/src/main/java/com/road2cs/services/SemestersService.java
@@ -208,7 +208,7 @@ public class SemestersService {
 
     }
 
-    private Semester createSemester(double number, String term, String... courseCodes) {
+    private Semester createSemester(int number, String term, String... courseCodes) {
         Semester semester = new Semester(number, term);
         for (String code : courseCodes) {
             semester.addCourse(courseService.getCourse(code));
